Sun, Jan 15, 2023, 3:14 PM +00:00

+++ hours/hours.c
@@ -28,7 +28,7 @@ int main(void)
 
 float calc_hours(int hours[], int weeks, char output)
 {
-    int total;
+    float total;
 
     for (int i = 0; i < weeks; i++)
     {
